refactor(bridge): extract forwarding state
