+++ whatsrust/lib/main_test.go
@@ -21,7 +21,6 @@ import (
 	"go.mau.fi/whatsmeow/proto/waWeb"
 	"go.mau.fi/whatsmeow/types"
 	"go.mau.fi/whatsmeow/types/events"
-	"google.golang.org/protobuf/proto"
 )
 
 func profilePicturePNG(t *testing.T) []byte {
@@ -1101,25 +1100,6 @@ func messageCaption(message *waE2E.Message) *string {
 		return message.DocumentMessage.Caption
 	default:
 		return nil
-	}
-}
-
-func TestForwardingStateExtractsEverySupportedMessageContext(t *testing.T) {
-	forwarded, score := proto.Bool(true), uint32(5)
-	context := &waE2E.ContextInfo{IsForwarded: forwarded, ForwardingScore: &score}
-	for _, message := range []*waE2E.Message{
-		{ExtendedTextMessage: &waE2E.ExtendedTextMessage{ContextInfo: context}},
-		{ImageMessage: &waE2E.ImageMessage{ContextInfo: context}}, {VideoMessage: &waE2E.VideoMessage{ContextInfo: context}},
-		{AudioMessage: &waE2E.AudioMessage{ContextInfo: context}}, {DocumentMessage: &waE2E.DocumentMessage{ContextInfo: context}},
-		{StickerMessage: &waE2E.StickerMessage{ContextInfo: context}},
-	} {
-		state := forwardingStateFromMessage(message)
-		if !state.isForwarded || state.score != 5 {
-			t.Fatalf("forwarding state = %#v", state)
-		}
-	}
-	if state := forwardingStateFromMessage(&waE2E.Message{Conversation: stringPointer("text")}); state != (forwardingState{}) {
-		t.Fatalf("conversation forwarding state = %#v", state)
 	}
 }
